refactor: RacingGameHost 클래스 와 Application 클래스 분리
